+++ my-app/src/components/App/App.jsx
@@ -1,37 +1,17 @@
-import React, { useCallback, useContext } from "react";
-import style from './style.module.scss';
+import React, { useCallback } from "react";
 import { useEffect, useState } from 'react';
-import axios from 'axios';
+import style from './style.module.scss';
 
 import Header from '../Header/Header';
 import Hero from '../Hero/Hero';
 import Movies from '../Movies/Movies';
 import MoviePopUp from "../MoviePopUp/MoviePopUp";
-import PopularMoviesContext from './PopularMoviesContext'
-
-// const fetchMoviesData = async (callback) => {
-//     const moviesData = await axios.get(`https://api.themoviedb.org/3/movie/popular?api_key=${process.env.REACT_APP_API_KEY}`);
-//     callback(moviesData.data.results);
-// }
-
-const moviesCategories = [{ title: 'Popular Movies', endpoint: 'popular' },
-{ title: 'Top Movies', endpoint: 'top_rated' },
-{ title: 'Upcoming Movies', endpoint: 'upcoming' }];
-
-const fetchMoviesData = async (callback) => {
-    let moviesData = [];
-
-    for (const category of moviesCategories) {
-        const response = await axios
-            .get(`https://api.themoviedb.org/3/movie/${category.endpoint}?api_key=${process.env.REACT_APP_API_KEY}`);
-        moviesData.push(Array.from(response.data.results));
-    }
-    callback(moviesData);
-}
+import PopularMoviesContext from '../Context/AllMoviesContext'
+import Preloader from "../Preloader/Preloader";
+import {fetchMoviesData, moviesCategories } from "../../moviesApi/fetchMoviesData";
 
 const App = () => {
 
-    // const [popularMovies, setPopularMovies] = useState(null);
     const [allMovies, setAllMovies] = useState([]);
 
     const [popup, setShowPopup] = useState({ show: false, movieId: '' });
@@ -41,36 +21,32 @@ const App = () => {
     }, [popup]);
 
     useEffect(() => {
-        // fetchMoviesData(setPopularMovies);
         fetchMoviesData(setAllMovies);
     }, []);
 
+    return (
+        allMovies.length <= 0 ? <Preloader /> :
+            <PopularMoviesContext.Provider value={allMovies[0]}>
 
-    // return popularMovies && (
-    return allMovies && (
+                <div className={style.app}>
+                    <div className={style.app__wrapper}>
+                        <div className={style.app__page}>
 
-        <div className={style.app}>
-            <div className={style.app__wrapper}>
-                <div className={style.app__page}>
-                    <Header />
-                    {/* <PopularMoviesContext.Provider value={popularMovies}> */}
-                    <PopularMoviesContext.Provider value={allMovies[0]}>
-                        <Hero />
-                    </PopularMoviesContext.Provider>
-                    {/* {
-                        moviesCategories.map((category, index) =>
-                            <Movies key={index} title={category.title} moviesEndpoint={category.endpoint} setShowPopup={getMovieData} />)
-                    } */}
-                    {
-                        allMovies.map((movies, index) =>
-                            <Movies key={index} title={moviesCategories[index].title} movies={movies}  setShowPopup={getMovieData} />)
-                    }
+                            <Header />
+                            <Hero />
+                            {
+                                allMovies.map((movies, index) =>
+                                    <Movies key={index} title={moviesCategories[index].title} movies={movies} setShowPopup={getMovieData} />)
+                            }
+                        </div>
+
+                        {
+                            popup.show && <MoviePopUp popupData={popup} setShowPopup={getMovieData}></MoviePopUp>
+                        }
+
+                    </div>
                 </div>
-                {
-                    popup.show && <MoviePopUp popupData={popup} setShowPopup={getMovieData}></MoviePopUp>
-                }
-            </div>
-        </div>
+            </PopularMoviesContext.Provider>
     )
 }
