+++ my-app/src/components/MoviePopUp/MoviePopUp.jsx
@@ -1,24 +1,17 @@
-import axios from 'axios';
 import { useEffect, useState } from 'react';
 import style from './style.module.scss';
 import { ReactComponent as CloseSvg } from '../../icons/close.svg';
+
 import MovieDescription from './MovieDescription';
 import MovieAdditional from './MovieAdditional';
+import fetchMovieDataInPopup from '../../moviesApi/fetchMovieDataInPopup';
 
 const MoviePopUp = ({ popupData, setShowPopup }) => {
 
     const [movie, setMovie] = useState(null);
 
     useEffect(() => {
-        const fetchMovieData = async () => {
-            const movieData = await axios.get(`https://api.themoviedb.org/3/movie/${popupData.movieId}?api_key=${process.env.REACT_APP_API_KEY}`);
-            const movieKey = await axios.get(`https://api.themoviedb.org/3/movie/${popupData.movieId}/videos?api_key=${process.env.REACT_APP_API_KEY}`);
-            if (movieKey.data.results.length > 0) {
-                movieData.data.key = movieKey.data.results[0].key;
-            }
-            setMovie(movieData.data);
-        }
-        fetchMovieData();
+        fetchMovieDataInPopup(popupData,setMovie);
     }, []);
 
 
@@ -32,9 +25,6 @@ const MoviePopUp = ({ popupData, setShowPopup }) => {
                 <div className={style.popup__right}>
                     <MovieDescription movie={movie} />
                 </div>
-                {/* <div>
-                    <iframe width="560" height="315" src={`https://www.youtube.com/embed/${movie.key}`} title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
-                </div> */}
             </div>
 
         </div>
