--- my-app/src/components/MoviePopUp/PopupMovieInfo.jsx
@@ -1,11 +1,0 @@
-import style from './style.module.scss';
-
-const PopupMovieInfo = ({ movie }) => {
-    return (
-        <div>
-            
-        </div>
-    )
-}
-
-export default PopupMovieInfo;
