+++ my-app/src/components/MoviePopUp/style.module.scss
@@ -6,7 +6,7 @@
     width: 100%;
     height: 100%;
     background-color: rgba(70, 70, 70, 0.1);
-    overflow: auto;
+    overflow-y: hidden;
     color: #fff;
     &__close {
         top: 20px;
