+++ my-app/src/components/Preloader/Preloader.jsx
@@ -1,0 +1,11 @@
+import CircularProgress from "react-cssfx-loading/lib/CircularProgress";
+import style from './style.module.scss'
+
+const Preloader = () => {
+    return (
+        <div className={style.preloader}>
+            <CircularProgress color="#FF0000" />
+        </div>
+    )
+}
+export default Preloader;
