+++ my-app/src/components/Preloader/style.module.scss
@@ -1,0 +1,9 @@
+.preloader{
+    width: 100%;
+    height: 100%;
+    svg{
+        width: 500px;
+        height: 500px;
+        margin: 0 auto;
+    }
+}
