+++ my-app/src/index.js
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom';
 import App from './components/App/App';
 import style from './index.module.scss';
 
-
 ReactDOM.render(
     <App />,
     document.getElementById('root')
