+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "react-cssfx-loading": "^1.0.3"
+  }
+}
